Turtlebot Control + ROS Skeleton Calc Script
This python script contains the code which controls the turtlebot, as well as handles the readings of the human skeleton by the Kinect. By using ROS Skeleton, the human joints are read as coordinates, using which vector calculation is performed. This script also contains the socket client code which sends the hand position values to the server running the AR.